FlightTasks: Add task for acceleration input by stick
This is a draft for a more intuitive position controlled flight by stick.
Mapping the stick input to velocity generally results in a too agressive
flight behavior so often tight acceleration limits have to be applied
which result in a too long braking distance and a delay in executing stick
inputs.

Mapping the stick inputs to an acceleration looks promising to solve a big
part of the listed problems. It feels more intuitive and guarantees gentle
vehicle behavior when being gentle on the sticks while allowing to brake
hard in emergency cases by pulling the stick all the way to the oposite
side of flight direction.

This task requires changes to the position controller to track an
acceleration setpoint which me and @bresch are working on.